requirements.txt
python-telegram-bot>=20.0
supabase>=1.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
openpyxl>=3.1.0

# Optional dependencies for full functionality
# These require system-level installations
# ffmpeg (for audio/video conversion)
# imagemagick (for SVG conversion)
# libreoffice (for document conversion)
# pandoc (for advanced document conversion)